% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterXts.R
\name{filterXts}
\alias{filterXts}
\title{Filter data.frame with xts syntax}
\usage{
filterXts(df, xtsstring, tzone = "UTC", by_colname = NULL)
}
\arguments{
\item{df}{a data.frame with at least one column of POSIXct values.}

\item{xtsstring}{a string which describes the desired date range to be
filtered out (see \code{\link[xts]{xts}}).}

\item{tzone}{give the timezone format to be used. default = "UTC"}

\item{by_colname}{give the column name of the datetime column to be used.}
}
\value{
a filtered data.frame which is a subset of the original data.frame \code{df}
Prepare a named character to be used in the by = statement in the join function.
Its name has to be the column name of the matching column from original

Creating a named Vector for inner_join its by argument
}
\description{
This function converts your data.frame which must have a cloumn with POSIXct
  values into an xts object and filters the xts object according to the
  argument xtsstring. It converts the filtered xts back into a data.frame
  and returns it.
}
